server/requirements.txt
Flask==3.0.3
flask-mysql-connector==1.1.0
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
python-dotenv==1.0.1
mysql-connector-python==9.3.0
PyMySQL==1.1.0
cryptography==42.0.5
Werkzeug==3.0.4
redis==5.0.1
python-json-logger==3.3.0
prometheus-flask-exporter==0.23.0
prometheus_client==0.21.1
psutil==7.0.0
python-decouple==3.8
PyYAML==6.0.1
